Move static methods out of SnapshotPolicySchedule into new class util.SnapshotPolicySchedules (and fix the FIXMEs and a missing license header at the same time)
